Convert Expression to typescript
